KOGITO-1830: BDD tests: Test service discovery on Spring Boot application
