+++ onboarding-example/README.md
@@ -38,7 +38,7 @@ Please follow the instruction for each of the individual services. It is recomen
 
 * [hr](hr/README.md)
 * [payroll](payroll/README.md)
-* [onboarding](onboarding/README.md)
+* [onboarding quarkus](onboarding-quarkus/README.md) or [onboarding spring boot](onboarding-springboot/README.md)
 
 ## Example Usage
