+++ onboarding-example/onboarding-quarkus/onboarding-dashboard.json
