+++ onboarding-example/onboarding-quarkus/reflection.json
