+++ onboarding-example/onboarding-quarkus/src/main/java/org/kie/kogito/examples/onboarding/Payroll.java
@@ -1,0 +1,68 @@
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.kie.kogito.examples.onboarding;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+/**
+ * This class was automatically generated by the data modeler tool.
+ */
+
+public class Payroll implements java.io.Serializable {
+
+	static final long serialVersionUID = 1L;
+
+	private java.lang.Integer vacationDays;
+	private java.lang.Double taxRate;
+	
+	@JsonFormat(shape = JsonFormat.Shape.STRING)
+	private java.util.Date paymentDate;
+
+	public Payroll() {
+	}
+
+	public java.lang.Integer getVacationDays() {
+		return this.vacationDays;
+	}
+
+	public void setVacationDays(java.lang.Integer vacationDays) {
+		this.vacationDays = vacationDays;
+	}
+
+	public java.lang.Double getTaxRate() {
+		return this.taxRate;
+	}
+
+	public void setTaxRate(java.lang.Double taxRate) {
+		this.taxRate = taxRate;
+	}
+
+	public java.util.Date getPaymentDate() {
+		return this.paymentDate;
+	}
+
+	public void setPaymentDate(java.util.Date paymentDate) {
+		this.paymentDate = paymentDate;
+	}
+
+	public Payroll(java.lang.Integer vacationDays, java.lang.Double taxRate,
+			java.util.Date paymentDate) {
+		this.vacationDays = vacationDays;
+		this.taxRate = taxRate;
+		this.paymentDate = paymentDate;
+	}
+
+}
