+++ onboarding-example/onboarding-quarkus/src/main/resources/META-INF/kmodule.xml
