+++ onboarding-example/onboarding-quarkus/src/main/resources/application.properties
