+++ onboarding-example/onboarding-quarkus/src/main/resources/org/kie/kogito/examples/onboarding/setupHR.bpmn
