+++ onboarding-example/onboarding-quarkus/src/test/resources/application.properties
