+++ onboarding-example/onboarding-springboot/src/main/resources/META-INF/kmodule.xml
@@ -1,0 +1,1 @@
+<kmodule xmlns="http://www.drools.org/xsd/kmodule" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"/>
