+++ onboarding-example/onboarding-springboot/src/main/resources/application.properties
@@ -1,0 +1,5 @@
+spring.mvc.servlet.path=/docs
+server.port=8080
+resteasy.jaxrs.scan-packages=org.kie.kogito.**,http*
+
+infinispan.remote.server-list=127.0.0.1:11222
