+++ onboarding-example/onboarding-springboot/src/test/java/org/kie/kogito/examples/test/RecordedOutputWorkItemHandler.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.kie.kogito.examples.test;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.function.Function;
+
+import org.kie.api.runtime.process.WorkItem;
+import org.kie.api.runtime.process.WorkItemHandler;
+import org.kie.api.runtime.process.WorkItemManager;
+
+public class RecordedOutputWorkItemHandler implements WorkItemHandler {
+    
+    private Map<String, Function<WorkItem, Map<String, Object>>> recorded = new HashMap<>();
+
+    @Override
+    public void executeWorkItem(WorkItem workItem, WorkItemManager manager) {
+        Map<String, Object> results = recorded.remove(workItem.getParameter("TaskName")).apply(workItem);
+
+        manager.completeWorkItem(workItem.getId(), results);
+    }
+
+    @Override
+    public void abortWorkItem(WorkItem workItem, WorkItemManager manager) {
+
+    }
+
+    public void record(String name, Function<WorkItem, Map<String, Object>> item) {
+        this.recorded.put(name, item);
+    }
+}
