--- onboarding-example/onboarding/.gitignore
@@ -1,34 +1,0 @@
-bin/
-/target
-/local
-
-# Eclipse, Netbeans and IntelliJ files
-/.*
-!.gitignore
-/nbproject
-/*.ipr
-/*.iws
-/*.iml
-
-# Repository wide ignore mac DS_Store files
-.DS_Store
-
-# Original jbpm ignores
-*~
-
-# Test info
-/settings*.xml
-/lib-jdbc/
-*.db
-*.tlog
-quarkus.log
-libsunec.*
-
-# modules that don't exist in this branch
-/jbpm-human-task-war/
-/jbpm-bam/
-/jbpm-gwt/
-
-# files used for external db testing
-jdbc_driver.jar
-db-settings.xml
